added heredoc detectors, started shortening preex markup
