+++ Makefile
@@ -44,7 +44,7 @@ COMMAND_EXPANSION_NAMES = substitute_envvars.c expand.c
 COMMAND_EXPANSION_F = command_expansion
 COMMAND_EXPANSION_SRCS = $(addprefix $(COMMAND_EXPANSION_F)/, $(COMMAND_EXPANSION_NAMES))
 
-EXECUTION_NAMES = exec_text_tree.c exec_text_tree_node.c exec_text_tree_controls.c redirections.c pipes.c files.c is_pipe.c is_redir.c is_redir_file.c navigation.c parent_n_child.c preexecution_markup.c
+EXECUTION_NAMES = exec_text_tree.c exec_text_tree_node.c exec_text_tree_controls.c redirections.c pipes.c files.c is_pipe.c is_redir.c is_redir_file.c navigation.c parent_n_child.c preexecution_markup.c is_node_interacting.c
 EXECUTION_F = execution
 EXECUTION_SRCS = $(addprefix $(EXECUTION_F)/, $(EXECUTION_NAMES))
 
@@ -197,7 +197,7 @@ vania:
 minivania:
 	$(PREFIX)make all_fancy test && ./$(TEST_FNAME)
 
-CMD = "cat < one.txt | grep 1 > two.txt"
+CMD = "cat << END | grep 1 > one.txt"
 run:
 	$(PREFIX)./minishell --debug 136 --command $(CMD)
 debug:
